Refactor to psutil, structurized project
