+++ pg_view/consts.py
@@ -1,0 +1,9 @@
+# some global variables for keyboard output
+freeze = False
+filter_aux = True
+autohide_fields = False
+display_units = False
+notrim = False
+realtime = False
+
+TICK_LENGTH = 1
